clean up of model, temp comment out current user update and create by model columns until auth created, can create and read from swagger

+++ src/main/java/com/rak/scheduler/service/AppointmentService.java
@@ -15,12 +15,16 @@ public class AppointmentService {
 	@Autowired
 	AppointmentRepository repository;
 
-	public void saveOrUpdates(Appointment appointment) {
-		repository.save(appointment);
+	public Appointment saveOrUpdates(Appointment appointment) {
+		return repository.save(appointment);
 	}
 
-	public void delete(long id) {
-		repository.deleteById(id);
+	public Optional<Appointment> delete(long id) {
+		Optional<Appointment> appt = this.getAppointmentById(id);
+		if(appt.isPresent()) {
+			repository.deleteById(id);
+		}
+		return appt;
 	}
 
 	public List<Appointment> getAppointments() {
